# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.15.0",
        "title": "Mocked-API",
        "description": "An API library for providing endpoints of mock data for use during testing. This API is still in the development and until it reaches version 1.0.0 the API is subject to change. To find out more head over to https://mocked-api.dev/"
    },
    "host": "api.mocked-api.dev",
    "schemes": [
        "https"
    ],
    "basePath": "/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
...
    "paths": {
...
        "/chat/random/{userCount}/{messageCount}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userCount",
                        "description": "The amount of users the messages can be generated from",
                        "type": "string",
                        "default": 100
                    },
                    {
                        "in": "path",
                        "name": "messageCount",
                        "description": "The amount of messages that will be returned",
                        "type": "string",
                        "default": 1000
                    }
                ],
                "summary": "Returns a random chat of messages from defined counts using randomly generated sentances",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MockChatResponse"
                        }
                    }
                }
            }
        },
...
        "/countries/{filterBy}": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Obtain a list of all countries that contain filtered text",
                "parameters": [
                    {
                        "in": "path",
                        "name": "filterBy",
                        "description": "The text you would like to filter countries by",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "United Kingdom"
                            }
                        }
                    }
                }
            }
        },
...
        "/currencies/digital-coins/bitcoin/random-address/{qty}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "qty",
                        "description": "The amount of addresses you require",
                        "type": "string",
                        "default": 10
                    }
                ],
                "summary": "Returns random bitcoin addresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/currencies/digital-coins/ethereum/random-address/{qty}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "qty",
                        "description": "The amount of addresses you require",
                        "type": "string",
                        "default": 10
                    }
                ],
                "summary": "Returns random ethereum addresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/currencies/digital-coins/litecoin/random-address/{qty}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "qty",
                        "description": "The amount of addresses you require",
                        "type": "string",
                        "default": 10
                    }
                ],
                "summary": "Returns random litecoin addresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/elements/{qty}": {
            "get": {
                "tags": [
                    "Chemical Elements"
                ],
                "summary": "Get a random list of value of Chemical Elements",
                "parameters": [
                    {
                        "in": "path",
                        "name": "qty",
                        "default": 10,
                        "description": "The amount of results you would like returned"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MockChemicalElements"
                        }
                    }
                }
            }
        },
...
    }
}